[tg] Move code from `handlers.py` to `get_ids.py`
